+++ inst/unit_tests/test_checks_2.R
@@ -77,5 +77,31 @@ test_checks_2 <- function() {
                            checkKeys = TRUE)
   })
 
+  # key collision
+  d <- wrapr::build_frame(
+    "id", "meas", "value" |
+    1   , "AUC" , 0.7     |
+    1   , "R2"  , NA      |
+    2   , "id" , 0.3      )
+  RUnit::checkException({
+  cdata::pivot_to_rowrecs(d,
+                          columnToTakeKeysFrom = 'meas',
+                          columnToTakeValuesFrom = 'value',
+                          rowKeyColumns = 'id')
+  })
+
+  # key collision
+  z <- wrapr::build_frame(
+    "meas", "AUC"   , "R2"     |
+      1   , 0.7     , NA_real_ |
+      2   , 0.5     , 0.3      )
+  RUnit::checkException({
+    cdata::unpivot_to_blocks(z,
+                             nameForNewKeyColumn = "meas",
+                             nameForNewValueColumn = "value",
+                             columnsToTakeFrom = c("AUC", "R2"),
+                             checkKeys = TRUE)
+  })
+
   invisible(NULL)
 }